Issue #1909 - remove deprecated Unsafe usage (requires at least jdk15)
