+++ pom.xml
@@ -276,7 +276,7 @@ See the Apache License Version 2.0 for the specific language governing permissio
           <artifactId>maven-compiler-plugin</artifactId>
           <version>3.11.0</version>
           <configuration>
-            <release>11</release>
+            <release>15</release>
             <compilerArgs>
               <arg>-parameters</arg>
             </compilerArgs>
